+++ R/content.R
@@ -313,19 +313,36 @@ tool_string <- function(x) {
 ContentJson <- new_class(
   "ContentJson",
   parent = Content,
-  properties = list(value = class_any)
+  properties = list(
+    # Some providers guarantee JSON returning it as part of the parsed results.
+    # Others return as a string, which may or may not be valid JSON. We don't
+    # want to force conversion on construction() since that's likely to be
+    # inconveniently early
+    data = class_any,
+    string = prop_string(allow_null = TRUE),
+
+    # Virtual property that always returns parsed JSON or errors
+    parsed = new_property(getter = function(self) {
+      if (is.null(self@string)) {
+        self@data
+      } else {
+        jsonlite::parse_json(self@string)
+      }
+    })
+  )
 )
+
 method(format, ContentJson) <- function(x, ...) {
   paste0(
     cli::format_inline("[{.strong data}] "),
-    pretty_json(x@value)
+    pretty_json(x@parsed)
   )
 }
 method(contents_html, ContentJson) <- function(content) {
-  sprintf('<pre><code>%s</code></pre>\n', pretty_json(content@value))
+  sprintf('<pre><code>%s</code></pre>\n', pretty_json(content@parsed))
 }
 method(contents_markdown, ContentJson) <- function(content) {
-  sprintf('```json\n%s\n```\n', pretty_json(content@value))
+  sprintf('```json\n%s\n```\n', pretty_json(content@parsed))
 }
 
 ContentUploaded <- new_class(